+++ src/stable_simd/mod.rs
@@ -1,7 +1,37 @@
-mod rng128;
-mod rng256;
-mod rng512;
+#[cfg(all(
+any(target_arch = "x86", target_arch = "x86_64"),
+target_feature = "avx2"
+))]
+mod avx2;
 
-pub use rng128::Rng128;
-pub use rng256::Rng256;
-pub use rng512::Rng512;
+#[cfg(all(
+any(target_arch = "x86", target_arch = "x86_64"),
+not(target_feature = "avx2"),
+target_feature = "sse2"
+))]
+mod sse2;
+
+#[cfg(not(all(
+any(target_arch = "x86", target_arch = "x86_64"),
+any(target_feature = "sse2", target_feature = "avx2"),
+)))]
+mod fallback;
+
+#[cfg(all(
+any(target_arch = "x86", target_arch = "x86_64"),
+target_feature = "avx2"
+))]
+pub use avx2::RngWide;
+
+#[cfg(all(
+any(target_arch = "x86", target_arch = "x86_64"),
+not(target_feature = "avx2"),
+target_feature = "sse2"
+))]
+pub use sse2::RngWide;
+
+#[cfg(not(all(
+any(target_arch = "x86", target_arch = "x86_64"),
+any(target_feature = "sse2", target_feature = "avx2"),
+)))]
+pub use fallback::RngWide;